requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cryptography==43.0.3
Flask==2.3.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
oauthlib==3.3.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
URLObject==3.0.0
Werkzeug==3.1.3
